Fix initial highlighted color scheme with SublimeLinter being installed
SublimeLinter is the only package, which creates custom color schemes to dynamically add support for its scopes.
It modifies the "color_scheme" setting each time a new color scheme without support for SublimeLinter is selected.

This causes heavy duplicates on the quick panel which are filtered out by the changes of the previous commit.

To keep initial_highlight point to the correct quick panel item, we need to find the original color scheme SublimeLinter manipulated.